fix(gui): restore React Fast Refresh and split documented-limits logic
Two dev/runtime fixes for the Providers rate-limits UI:

- Upgrade @vitejs/plugin-react 6.0.3 -> 6.0.5. The 6.0.3 release failed to
  inject the fast-refresh preamble in dev on Vite 8, so every component
  module threw `Uncaught ReferenceError: $RefreshReg$ is not defined` and
  the dashboard would not boot under `bun run dev`.

- Split ProviderDocumentedLimits.tsx into a component-only module plus a
  pure provider-workspace/documented-limits.ts (types, FREE_TIER_KEYS,
  formatDocumentedLimits). The component file previously exported
  non-component helpers, which violates react-refresh/only-export-components
  and breaks fast refresh; the lint gate flagged it.
